utils and view refactoring
2021/08/23
- created utils submodule and updated docs accordingly
- minimal changes to ease syntax on imports
- updated readme.md
- renamed _InactiveContainer to InactiveContainer
- renamed _WaitContainer to InactiveContainer
