+++ docs/threadmonitor.controller.rst
@@ -1,17 +1,6 @@
 threadmonitor.controller package
 ================================
 
-Submodules
-----------
-
-threadmonitor.controller.controller module
-------------------------------------------
-
-.. automodule:: threadmonitor.controller.controller
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 Module contents
 ---------------
