+++ threadmonitor/model/__init__.py
@@ -1,3 +1,23 @@
 """
 TODO Insert model module description here
 """
+
+### INIZIALIZZAZIONE LOGICA ###
+
+### hashMap contenente come chiave il nome del lock, e come valore il relativo canvas ###
+#self.lockContainer = {}
+
+### In quale container è conenuto al momento il thread ###
+#self.threadContainer = {}
+
+### wait container associato a uno specifico lock ###
+#self.waitContainer = {}
+
+### Lista dei threads ###
+#self.threads = []
+
+### Lista dei thread inattivi ###
+#self.inactiveThread = []
+
+### Lista dei thread in wait su un determinato lock ###
+#self.waitThread = {}
