--- threadmonitor/model/logic.py
@@ -1,19 +1,0 @@
-### INIZIALIZZAZIONE LOGICA ###
-
-### hashMap contenente come chiave il nome del lock, e come valore il relativo canvas ###
-#self.lockContainer = {}
-
-### In quale container è conenuto al momento il thread ###
-#self.threadContainer = {}
-
-### wait container associato a uno specifico lock ###
-#self.waitContainer = {}
-
-### Lista dei threads ###
-#self.threads = []
-
-### Lista dei thread inattivi ###
-#self.inactiveThread = []
-
-### Lista dei thread in wait su un determinato lock ###
-#self.waitThread = {}
